requirements.txt
alembic==1.4.3
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
cryptography==3.1
ecdsa==0.14.1; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
fastapi==0.61.1
geojson==2.5.0
h11==0.9.0
httpcore==0.10.2; python_version >= '3.6'
httptools==0.1.1; sys_platform != 'win32' and sys_platform != 'cygwin' and platform_python_implementation != 'PyPy'
httpx==0.14.3
idna==2.10
itsdangerous==1.1.0
mako==1.1.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
markupsafe==1.1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
mysqlclient==2.0.1
polyline==1.4.0
pyasn1==0.4.8
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pydantic==1.6.1; python_version >= '3.6'
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-editor==1.0.4
python-jose==3.2.0
rfc3986[idna2008]==1.4.0
rsa==4.6; python_version >= '3.5' and python_version < '4'
shortuuid==1.0.1
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.1.0; python_version >= '3.5'
sqlalchemy==1.3.19
starlette==0.13.6; python_version >= '3.6'
ujson==3.2.0
uvicorn==0.11.8
uvloop==0.14.0; sys_platform != 'win32' and sys_platform != 'cygwin' and platform_python_implementation != 'PyPy'
websockets==8.1; python_full_version >= '3.6.1'
